refactor(bflp): resolve all 40 SonarQube issues
Quality gate already green; cleanup brings the open issue count to zero.

Highlights:
- Cognitive complexity: extracted helpers in AuditLogger.list() (26→<15)
  and SettingsService.saveGlobalSettings() (34→<15).
- Builder pattern for value objects with too many constructor params:
  BanContext, GlobalSettings, FailureEvent. SaveGlobalSettings now takes
  a GlobalSettingsUpdate parameter object.
- `record` is a reserved word in Java 16+: renamed SPI method
  FailureRecorder.record() → recordEvent(); updated implementers and
  tests.
- HazelcastInstanceManager: dropped the unused clusterManager field and
  the dead `discoveryServices != null` check; OSGi binding kept as a
  no-op setter for DS wiring.
- Various: removed unused imports, redundant casts, unguarded logger
  calls, deprecated Proxy.getProxyClass usage (suppressed with
  justification — it is the contract API for
  ObjectInputStream.resolveProxyClass).

Suppressions (with reason):
- java:S1075 on Constants — JCR paths use '/' by spec
- java:S1874 in ClassLoaderAwareObjectInputStream — Proxy.getProxyClass
  required by resolveProxyClass contract
- java:S107 on the GraphQL saveGlobalSettings mutation — argument names
  form the public GraphQL schema and cannot be collapsed

Tests: 34/34 pass. SonarQube post-scan: 0 open issues.

+++ src/main/java/org/jahia/modules/bruteforceloginprotection/core/RecidiveCalculator.java
@@ -18,7 +18,7 @@ public final class RecidiveCalculator {
                 return capSec;
             }
         }
-        long result = (long) Math.min(duration, (double) Long.MAX_VALUE);
+        long result = (long) Math.min(duration, Long.MAX_VALUE);
         if (capSec > 0 && result > capSec) {
             return capSec;
         }
